Expose Streaming runtime
Combine stream and processor command under the streaming command.
