--- docs/riff_processor.md
@@ -1,35 +1,0 @@
----
-id: riff-processor
-title: "riff processor"
----
-## riff processor
-
-(experimental) processors apply functions to messages on streams
-
-### Synopsis
-
-<todo>
-
-### Options
-
-```
-  -h, --help   help for processor
-```
-
-### Options inherited from parent commands
-
-```
-      --config file        config file (default is $HOME/.riff.yaml)
-      --kube-config file   kubectl config file (default is $HOME/.kube/config)
-      --no-color           disable color output in terminals
-```
-
-### SEE ALSO
-
-* [riff](riff.md)	 - riff is for functions
-* [riff processor create](riff_processor_create.md)	 - create a processor to apply a function to messages on streams
-* [riff processor delete](riff_processor_delete.md)	 - delete processor(s)
-* [riff processor list](riff_processor_list.md)	 - table listing of processors
-* [riff processor status](riff_processor_status.md)	 - show processor status
-* [riff processor tail](riff_processor_tail.md)	 - watch processor logs
-
