--- docs/riff_stream.md
@@ -1,34 +1,0 @@
----
-id: riff-stream
-title: "riff stream"
----
-## riff stream
-
-(experimental) streams of messages
-
-### Synopsis
-
-<todo>
-
-### Options
-
-```
-  -h, --help   help for stream
-```
-
-### Options inherited from parent commands
-
-```
-      --config file        config file (default is $HOME/.riff.yaml)
-      --kube-config file   kubectl config file (default is $HOME/.kube/config)
-      --no-color           disable color output in terminals
-```
-
-### SEE ALSO
-
-* [riff](riff.md)	 - riff is for functions
-* [riff stream create](riff_stream_create.md)	 - create a stream of messages
-* [riff stream delete](riff_stream_delete.md)	 - delete stream(s)
-* [riff stream list](riff_stream_list.md)	 - table listing of streams
-* [riff stream status](riff_stream_status.md)	 - show stream status
-
