+++ docs/riff_streaming.md
@@ -1,0 +1,32 @@
+---
+id: riff-streaming
+title: "riff streaming"
+---
+## riff streaming
+
+(experimental) streaming runtime for riff functions
+
+### Synopsis
+
+<todo>
+
+### Options
+
+```
+  -h, --help   help for streaming
+```
+
+### Options inherited from parent commands
+
+```
+      --config file        config file (default is $HOME/.riff.yaml)
+      --kube-config file   kubectl config file (default is $HOME/.kube/config)
+      --no-color           disable color output in terminals
+```
+
+### SEE ALSO
+
+* [riff](riff.md)	 - riff is for functions
+* [riff streaming processor](riff_streaming_processor.md)	 - (experimental) processors apply functions to messages on streams
+* [riff streaming stream](riff_streaming_stream.md)	 - (experimental) streams of messages
+
