+++ docs/riff_streaming_stream.md
@@ -1,0 +1,34 @@
+---
+id: riff-streaming-stream
+title: "riff streaming stream"
+---
+## riff streaming stream
+
+(experimental) streams of messages
+
+### Synopsis
+
+<todo>
+
+### Options
+
+```
+  -h, --help   help for stream
+```
+
+### Options inherited from parent commands
+
+```
+      --config file        config file (default is $HOME/.riff.yaml)
+      --kube-config file   kubectl config file (default is $HOME/.kube/config)
+      --no-color           disable color output in terminals
+```
+
+### SEE ALSO
+
+* [riff streaming](riff_streaming.md)	 - (experimental) streaming runtime for riff functions
+* [riff streaming stream create](riff_streaming_stream_create.md)	 - create a stream of messages
+* [riff streaming stream delete](riff_streaming_stream_delete.md)	 - delete stream(s)
+* [riff streaming stream list](riff_streaming_stream_list.md)	 - table listing of streams
+* [riff streaming stream status](riff_streaming_stream_status.md)	 - show stream status
+
