+++ pkg/riff/commands/riff.go
@@ -54,8 +54,6 @@ map those streams to function inputs and outputs with processors.
 	cmd.AddCommand(NewContainerCommand(ctx, c))
 	cmd.AddCommand(NewFunctionCommand(ctx, c))
 	cmd.AddCommand(NewHandlerCommand(ctx, c))
-	cmd.AddCommand(NewStreamCommand(ctx, c))
-	cmd.AddCommand(NewProcessorCommand(ctx, c))
 
 	return cmd
 }
