+++ pkg/riff/commands/root.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/projectriff/cli/pkg/cli"
+	streamingcommands "github.com/projectriff/cli/pkg/streaming/commands"
 	"github.com/spf13/cobra"
 )
 
@@ -43,6 +44,9 @@ func NewRootCommand(ctx context.Context, c *cli.Config) *cobra.Command {
 	cmd.PersistentFlags().StringVar(&c.ViperConfigFile, cli.StripDash(cli.ConfigFlagName), "", fmt.Sprintf("config `file` (default is $HOME/.%s.yaml)", c.Name))
 	cmd.PersistentFlags().StringVar(&c.KubeConfigFile, cli.StripDash(cli.KubeConfigFlagName), "", "kubectl config `file` (default is $HOME/.kube/config)")
 	cmd.PersistentFlags().BoolVar(&color.NoColor, cli.StripDash(cli.NoColorFlagName), color.NoColor, "disable color output in terminals")
+
+	// add runtimes
+	cmd.AddCommand(streamingcommands.NewStreamingCommand(ctx, c))
 
 	// add root-only commands
 	cmd.AddCommand(NewCompletionCommand(ctx, c))
