+++ pkg/streaming/commands/processor.go
