+++ pkg/streaming/commands/processor_create.go
