+++ pkg/streaming/commands/processor_delete.go
