+++ pkg/streaming/commands/processor_tail.go
