+++ pkg/streaming/commands/stream.go
