+++ pkg/streaming/commands/stream_create.go
