+++ pkg/streaming/commands/stream_delete.go
