+++ pkg/streaming/commands/stream_status.go
