+++ pkg/streaming/commands/streaming_test.go
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package commands_test
+
+import (
+	"testing"
+
+	"github.com/projectriff/cli/pkg/streaming/commands"
+	rifftesting "github.com/projectriff/cli/pkg/testing"
+)
+
+func TestStreamingCommand(t *testing.T) {
+	table := rifftesting.CommandTable{
+		{
+			Name: "empty",
+			Args: []string{},
+		},
+	}
+
+	table.Run(t, commands.NewStreamingCommand)
+}
